pyproject.toml
[project]
name = "vinea"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "José de Jesus Filho", email = "jjesusfilho@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.3.2",
    "polars>=1.32.3",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "zeep[xmlsec]>=4.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
